requirements.txt
flask>=0.12.0
py-solc>=3.2.0
requests>=2.18.0
web3>=4.8.0
